Allow for multiple jobs to be created
